requirements.txt
pyzmq
simplejson
